:sparkles: HttpStartLine 구현 - StartLine 을 따로 분리하여 관리하기 위한 클래스 구현
